Hash: Merge sort_by_frequency into top_k_most_frequent_numbers
